a2a_customer_routing/requirements.txt
a2a-sdk==0.3.0
google-adk==1.11.0
google-genai==1.19.0
httpx==0.28.1
litellm==1.72.2
llama-index-core==0.12.40
llama-index-embeddings-nebius==0.3.1
llama-index-llms-nebius==0.1.2
nest-asyncio==1.6.0      
python-dotenv==1.1.1
requests==2.32.3     
streamlit==1.37.1    
uvicorn==0.34.2
